change id to int64

+++ vector_quants/quantizers/finite_scalar_quantizer.py
@@ -15,10 +15,10 @@ class FiniteScalarQuantizer(BaseVectorQuantizer):
         levels = cfg.levels
         # get params end
 
-        _levels = torch.tensor(levels, dtype=torch.int32)
+        _levels = torch.tensor(levels, dtype=torch.int64)
         self.register_buffer("_levels", _levels, persistent=False)
         _basis = torch.cumprod(
-            torch.tensor([1] + levels[:-1]), dim=0, dtype=torch.int32
+            torch.tensor([1] + levels[:-1]), dim=0, dtype=torch.int64
         )
         self.register_buffer("_basis", _basis, persistent=False)
 
@@ -54,7 +54,7 @@ class FiniteScalarQuantizer(BaseVectorQuantizer):
         d = self._levels - 1
         number = round_ste(F.sigmoid(latent) * d)
         code = number / d
-        indice = (number * self._basis).sum(dim=-1).to(torch.int32)
+        indice = (number * self._basis).sum(dim=-1).to(torch.int64)
 
         return code, indice
 
@@ -62,8 +62,7 @@ class FiniteScalarQuantizer(BaseVectorQuantizer):
         # (b, *, d) -> (n, d)
         latent, ps = pack_one(latent, "* d")
         number = round_ste(F.sigmoid(latent) * (self._levels - 1))
-        indice = (number * self._basis).sum(dim=-1).to(torch.int32)
-
+        indice = (number * self._basis).sum(dim=-1).to(torch.int64)
         indice = unpack_one(indice, ps, "*")
 
         return indice
